refactor : CrewQueryService 도입
