+++ src/main/java/runnershigh/capstone/crew/service/CrewService.java
@@ -1,32 +1,21 @@
 package runnershigh.capstone.crew.service;
 
-import java.util.List;
-import java.util.Set;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
 import runnershigh.capstone.chat.domain.ChatRoom;
 import runnershigh.capstone.crew.domain.Crew;
 import runnershigh.capstone.crew.dto.request.CrewCreateRequest;
-import runnershigh.capstone.crew.dto.request.CrewSearchRequest;
 import runnershigh.capstone.crew.dto.request.CrewUpdateRequest;
 import runnershigh.capstone.crew.dto.response.CrewCreateResponse;
 import runnershigh.capstone.crew.dto.response.CrewDeleteResponse;
-import runnershigh.capstone.crew.dto.response.CrewDetailResponse;
-import runnershigh.capstone.crew.dto.response.CrewParticipantsDetailsResponse;
-import runnershigh.capstone.crew.dto.response.CrewSearchPagingResponse;
-import runnershigh.capstone.crew.dto.response.CrewSimpleResponse;
 import runnershigh.capstone.crew.dto.response.CrewUpdateResponse;
-import runnershigh.capstone.crew.enums.CrewUserRole;
 import runnershigh.capstone.crew.exception.CrewNotFoundException;
 import runnershigh.capstone.crew.repository.CrewRepository;
 import runnershigh.capstone.crew.service.mapper.CrewMapper;
 import runnershigh.capstone.crewparticipant.domain.CrewParticipant;
-import runnershigh.capstone.crewparticipant.service.CrewParticipantService;
 import runnershigh.capstone.crewscore.service.CrewScoreService;
 import runnershigh.capstone.geocoding.dto.FormattedAddressResponse;
 import runnershigh.capstone.geocoding.service.GeocodingService;
@@ -40,20 +29,20 @@ import runnershigh.capstone.user.service.UserService;
 @Service
 @AllArgsConstructor
 @Slf4j
+@Transactional
 public class CrewService {
 
-    private final CrewRepository crewRepository;
     private final CrewMapper crewMapper;
+
+    private final CrewRepository crewRepository;
 
     private final UserService userService;
     private final GeocodingService geocodingService;
     private final CrewScoreService crewScoreService;
     private final S3Service s3Service;
-    private final CrewParticipantService crewParticipantService;
 
     private static final String S3_DIRECTORY_NAME = "crew";
 
-    @Transactional
     public CrewCreateResponse createCrew(Long crewLeaderId, CrewCreateRequest crewCreateRequest,
         MultipartFile image) {
 
@@ -74,22 +63,6 @@ public class CrewService {
         return new CrewCreateResponse(crew.getId());
     }
 
-    @Transactional(readOnly = true)
-    public CrewDetailResponse getCrewDetail(Long userId, Long crewId) {
-        Crew crew = getCrewById(crewId);
-        crew.saveCrewRank(crewScoreService.getCrewRank(crewId));
-        CrewUserRole userRole = crew.validateAndReturnUserRole(userId);
-        Double score = crewScoreService.getCrewScore(crewId).getScore();
-        return crewMapper.toCrewDetailResponse(crew, userRole, score);
-    }
-
-    @Transactional(readOnly = true)
-    public Set<CrewParticipantsDetailsResponse> getCrewParticipants(Long crewId) {
-        Crew crew = getCrewByIdWithParticipants(crewId);
-        return crewMapper.toCrewParticipantsDetailsResponse(crew.getCrewParticipant());
-    }
-
-    @Transactional
     public CrewUpdateResponse updateCrew(Long crewLeaderId, CrewUpdateRequest crewUpdateRequest,
         Long crewId, MultipartFile image) {
         Crew crew = getCrewByIdAndLeaderId(crewId, crewLeaderId);
@@ -107,7 +80,6 @@ public class CrewService {
         return new CrewUpdateResponse(crew.getId());
     }
 
-    @Transactional
     public CrewDeleteResponse deleteCrew(Long crewLeaderId, Long crewId) {
         Crew crew = getCrewByIdAndLeaderId(crewId, crewLeaderId);
         s3Service.delete(crew.getImage());
@@ -116,61 +88,8 @@ public class CrewService {
         return new CrewDeleteResponse(crew.getId());
     }
 
-    @Transactional
-    public CrewSearchPagingResponse<CrewSimpleResponse> searchCrew(
-        CrewSearchRequest crewSearchRequest,
-        Pageable pageable) {
-
-        if (crewSearchRequest.name() != null) {
-            Page<Crew> crews = crewRepository.searchByName(crewSearchRequest.name(), pageable);
-            Page<CrewSimpleResponse> crewSimpleResponses = crewMapper.toCrewSimplePagingResponse(
-                crews);
-            return CrewSearchPagingResponse.from(crewSimpleResponses);
-
-        } else if (crewSearchRequest.region() != null) {
-            Page<Crew> crews = crewRepository.searchBySpecificLocation(crewSearchRequest.region(),
-                pageable);
-            Page<CrewSimpleResponse> crewSimpleResponses = crewMapper.toCrewSimplePagingResponse(
-                crews);
-            return CrewSearchPagingResponse.from(crewSimpleResponses);
-        }
-
-        return null;
-    }
-
-    @Transactional(readOnly = true)
-    public CrewSearchPagingResponse<CrewSimpleResponse> getCrewNearby(Long userId,
-        Pageable pageable) {
-
-        User user = userService.getUser(userId);
-        String city = user.getUserLocation().getCity();
-        String dong = user.getUserLocation().getDong();
-
-        Page<Crew> crews = crewRepository.findNearCrewWithoutParticipation(city, dong, userId,
-            pageable);
-
-        Page<CrewSimpleResponse> crewNearbyResponses = crewMapper.toCrewSimplePagingResponse(crews);
-
-        return CrewSearchPagingResponse.from(crewNearbyResponses);
-    }
-
-    public Crew getCrewById(Long crewId) {
-        return crewRepository.findById(crewId)
-            .orElseThrow(() -> new CrewNotFoundException(ErrorCode.CREW_NOT_FOUND));
-    }
-
-    @Transactional(readOnly = true)
-    public List<Crew> getCrewsByUserId(Long userId) {
-        return crewParticipantService.getCrewsByUserId(userId);
-    }
-
     private Crew getCrewByIdAndLeaderId(Long crewId, Long crewLeaderId) {
         return crewRepository.findByIdAndCrewLeaderId(crewId, crewLeaderId)
-            .orElseThrow(() -> new CrewNotFoundException(ErrorCode.CREW_NOT_FOUND));
-    }
-
-    private Crew getCrewByIdWithParticipants(Long crewId) {
-        return crewRepository.findByIdWithParticipants(crewId)
             .orElseThrow(() -> new CrewNotFoundException(ErrorCode.CREW_NOT_FOUND));
     }
